+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+[Unreleased]
+---------------------
+##### Added
+- Ability to pass an XForm preprocessor function.
+- Output a languageMap property to the result.
+
 [1.13.0] - 2016-12-20
 ---------------------
 ##### Added